pyproject.toml
[tool.poetry]
name = "target-airtable"
version = "0.0.1"
description = "`target-airtable` is a Singer target for Airtable, built with the Meltano SDK for Singer Targets."
authors = ["Tomas Votava"]
keywords = ["ELT", "Airtable"]
license = "MIT"
include = ["py.typed"]
repository = "https://github.com/tomasvotava/target-airtable"

[tool.poetry.dependencies]
python = "<3.11,>=3.8"
singer-sdk = "^0.38.0"

httpx = "^0.27.0"
[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
ruff = "^0.4.10"
mypy = "^1.10.0"
commitizen = "^3.27.0"

python-dotenv = "^1.0.1"
[build-system]
requires = ["poetry-core>=1.0.8"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
target-airtable = 'target_airtable.target:TargetAirtable.cli'

[tool.ruff]
target-version = "py39"
line-length = 120

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "B",
    "I",
    "N",
    "UP",
    "S",
    "A",
    "Q",
    "DTZ",
    "PT",
    "SIM",
    "PTH",
    "RUF",
    "T20",
    "ASYNC",
    "C4",
    "DJ",
    "ISC",
    "TCH",
    "TD",
]

[tool.ruff.lint.isort]
known-first-party = ["target_airtable"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = [
    "S101",   # allow asserts in tests
    "RUF018", # allow asserts with walrus in tests
    "D",      # ignore missing documentation in tests
]

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "semver"
version_provider = "poetry"
update_changelog_on_bump = true

[tool.mypy]
strict = true

[[tool.mypy.overrides]]
module = ["backoff.*"]
ignore_missing_imports = true
